+++ server/impl/src/main/java/com/walmartlabs/concord/server/security/internal/InternalRealm.java
@@ -1,0 +1,34 @@
+package com.walmartlabs.concord.server.security.internal;
+
+import com.walmartlabs.concord.server.security.PrincipalUtils;
+import com.walmartlabs.concord.server.security.UserPrincipal;
+import org.apache.shiro.authc.AuthenticationException;
+import org.apache.shiro.authc.AuthenticationInfo;
+import org.apache.shiro.authc.AuthenticationToken;
+import org.apache.shiro.authz.AuthorizationInfo;
+import org.apache.shiro.realm.AuthorizingRealm;
+import org.apache.shiro.subject.PrincipalCollection;
+
+import javax.inject.Named;
+
+@Named
+public class InternalRealm extends AuthorizingRealm {
+
+    public static final String REALM_NAME = "internal";
+
+    @Override
+    protected AuthenticationInfo doGetAuthenticationInfo(AuthenticationToken token) throws AuthenticationException {
+        // we don't need the authentication part
+        throw new IllegalStateException("Not implemented");
+    }
+
+    @Override
+    protected AuthorizationInfo doGetAuthorizationInfo(PrincipalCollection principals) {
+        UserPrincipal p = principals.oneByType(UserPrincipal.class);
+        if (!REALM_NAME.equals(p.getRealm())) {
+            return null;
+        }
+
+        return PrincipalUtils.toAuthorizationInfo(principals);
+    }
+}